# todo_rest_specs/todo_swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is the contract for our ToDo app. Data will be stored in Domino. The UI will be React",
    "version": "1.0.0",
    "title": "Swagger Domino/React ToDos",
    "contact": {
      "name": "Paul Withers",
      "email": "watsonwork.me/pwithers@intec.co.uk",
      "url": "http://www.intec.co.uk"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost:8081",
  "basePath": "/api",
...
  "schemes": [
    "http"
  ],
  "paths": {
    "/stores": {
      "get": {
        "tags": [
          "store"
        ],
        "summary": "Get all stores the user has access to",
        "description": "",
        "operationId": "getStores",
        "consumes": [
          "application/json; utf-8"
        ],
        "produces": [
          "application/json; utf-8"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Store"
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          }
        },
        "security": [
          {
            "domino_auth": []
          },
          {
            "application_auth": [],
            "user_auth": []
          }
        ]
      }
    },
    "/stores/todos/markOverdue": {
      "get": {
        "tags": [
          "store",
          "todo"
        ],
        "summary": "Mark any todos that are due as \"Overdue\" status",
        "description": "",
        "operationId": "markOverdue",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "202": {
            "description": "accepted for processing"
          },
          "401": {
            "description": "unauthenticated"
          }
        },
        "security": [
          {
            "domino_auth": [],
            "sched_auth": []
          },
          {
            "application_auth": [],
            "sched_auth": []
          }
        ]
      }
    },
...
    "/store/{store}": {
      "put": {
        "tags": [
          "store"
        ],
        "description": "Change the title of a ToDo store",
        "operationId": "changeTitle",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "store",
            "description": "store id or name",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "in": "query",
            "name": "newTitle",
            "description": "New title for the store",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/definitions/Store"
          },
          "400": {
            "description": "Invalid store name or id supplied"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "405": {
            "description": "Validation exception"
          }
        },
        "security": [
          {
            "domino_auth": []
          },
          {
            "application_auth": [],
            "user_auth": []
          }
        ]
      },
      "get": {
        "tags": [
          "store"
        ],
        "description": "Retrieve information about a ToDo store",
        "operationId": "getStoreInfo",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "store",
            "description": "store id or name",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/definitions/Store"
          },
          "400": {
            "description": "Invalid store name or id supplied"
          },
          "401": {
            "description": "Unauthenticated"
          }
        },
        "security": [
          {
            "domino_auth": []
          },
          {
            "application_auth": [],
            "user_auth": []
          }
        ]
      }
    },
...
    "/todo/{key}": {
      "post": {
        "tags": [
          "todo"
        ],
        "description": "add a todo to the store",
        "operationId": "addToDo",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "key",
            "description": "store id or name in which to post the ToDo",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "ToDo to post to the store",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ToDo"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returned ToDo",
            "schema": {
              "$ref": "#/definitions/ToDo"
            }
          },
          "400": {
            "description": "Invalid store name or id supplied"
          },
          "401": {
            "description": "Not authorised"
          }
        },
        "security": [
          {
            "domino_auth": []
          },
          {
            "application_auth": [],
            "user_auth": []
          }
        ]
      },
      "put": {
        "tags": [
          "todo"
        ],
        "description": "update a ToDo in the store",
        "operationId": "updateToDo",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "key",
            "description": "metaversalId of the ToDo",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "ToDo to update in the store",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ToDo"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returned ToDo",
            "schema": {
              "$ref": "#/definitions/ToDo"
            }
          },
          "400": {
            "description": "Invalid store name or id supplied"
          },
          "401": {
            "description": "Not authorised"
          }
        },
        "security": [
          {
            "domino_auth": []
          },
          {
            "application_auth": [],
            "user_auth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "todo"
        ],
        "description": "delete a ToDo from the store",
        "operationId": "deleteToDo",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "key",
            "description": "metaversalId of the ToDo",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "ToDo to update in the store",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ToDo"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "boolean success",
            "schema": {
              "type": "boolean",
              "description": "boolean success or failure"
            }
          },
          "400": {
            "description": "Invalid store name or id supplied"
          },
          "401": {
            "description": "Not authorised"
          }
        },
        "security": [
          {
            "domino_auth": []
          },
          {
            "application_auth": [],
            "user_auth": []
          }
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "domino_auth": {
      "type": "basic"
    },
    "application_auth": {
      "type": "apiKey",
      "in": "header",
      "name": "X-TODO-API-KEY"
    },
    "user_auth": {
      "type": "apiKey",
      "in": "header",
      "name": "X-TODO-USER-KEY"
    },
    "debug_auth": {
      "type": "apiKey",
      "in": "header",
      "name": "X-DEBUG-KEY"
    },
    "sched_auth": {
      "type": "apiKey",
      "in": "header",
      "name": "X-SCHED-KEY"
    }
  },
  "definitions": {
    "Store": {
      "type": "object",
      "properties": {
        "replicaId": {
          "type": "string",
          "description": "16 character hexadecimal string corresponding to NSF's replica id",
          "minLength": 16,
          "maxLength": 16,
          "format": "hex",
          "readOnly": true
        },
        "title": {
          "type": "string",
          "description": "Display name for store, e.g. a project name or an individual user for personal ToDos"
        },
        "name": {
          "type": "string",
          "description": "File name for the NSF, appended to the folder name where all ToDo stored are stored"
        },
        "type": {
          "type": "string",
          "description": "ToDo type, see enum'",
          "enum": [
            "Personal",
            "Team"
          ]
        }
      },
      "required": [
        "title",
        "name",
        "type"
      ]
    },
    "ToDo": {
      "type": "object",
      "properties": {
        "unid": {
          "type": "string",
          "description": "MetaversalID to access this ToDo from just a Domino Session. First 16 chars are replicaId, rest is UNID",
          "minLength": 48,
          "maxLength": 48,
          "format": "hex",
          "readOnly": true
        },
        "author": {
          "type": "string",
          "description": "username of the person who creates the ToDo - set automatically via the API from the currentUser passed across",
          "readOnly": true
        },
        "currentUser": {
          "$ref": "#/definitions/User"
        },
        "taskName": {
          "type": "string",
          "description": "Brief description of task"
        },
        "description": {
          "type": "string",
          "description": "Detailed description of the task"
        },
        "dueDate": {
          "type": "string",
          "format": "date",
          "description": "Date when the task is due in RFC 3339 section 5.6 format 2018-03-19"
        },
        "priority": {
          "type": "string",
          "description": "Priority for the ToDo, see enum",
          "enum": [
            "Low",
            "Medium",
            "High"
          ]
        },
        "assignedTo": {
          "type": "string",
          "description": "Name of the person the ToDo is assigned to. If not set, current user will be used"
        },
        "status": {
          "type": "string",
          "description": "Current status of the ToDo. Set via workflow actions, defaulting to New",
          "enum": [
            "New",
            "Reassigned",
            "Complete",
            "Overdue"
          ],
          "readOnly": true
        }
      },
      "required": [
        "currentUser",
        "taskName",
        "description",
        "dueDate",
        "priority"
      ]
    },
    "User": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string",
          "description": "If Domino authentication is in operation for REST access, this property will be ignored for the CurrentUser. If not, the REST service  will use the username passed, replacing '@' with '_', append a computed OU, and use the server's O."
        },
        "access": {
          "type": "object",
          "description": "Required when setting access levels, returned when checking access levels, irrelevant otherwise.",
          "properties": {
            "level": {
              "type": "string",
              "description": "Application access level. This will be converted to Editor or Reader in Domino ACL",
              "enum": [
                "Admin",
                "Reader",
                "Editor",
                "No Access"
              ]
            },
            "allowDelete": {
              "type": "boolean",
              "description": "Whether or not Delete documents privilege is added"
            }
          }
        }
      }
    }
  }
}